# 0627/go/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for managing books.",
        "title": "Book Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/books": {
            "get": {
                "description": "搜索书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书"
                ],
                "summary": "搜索书籍",
                "parameters": [
                    {
                        "description": "书籍信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.Book"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "book not found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "传入书籍信息新增书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书"
                ],
                "summary": "添加书籍",
                "parameters": [
                    {
                        "description": "书籍信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "book already exists",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "400错误",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "put": {
                "description": "更新书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书"
                ],
                "summary": "更新书籍",
                "parameters": [
                    {
                        "type": "string",
                        "description": "书籍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.Book"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "400错误",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书"
                ],
                "summary": "删除书籍",
                "parameters": [
                    {
                        "type": "string",
                        "description": "书籍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Book not exists",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "stock": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}